packages/backend/pyproject.toml
[project]
name = "loyalty-backend"
version = "0.1.0"
description = "Loyalty Platform API (FastAPI + DynamoDB)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.35.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "mangum>=0.17.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "httpx>=0.27.0",
    "moto[dynamodb]>=5.0",
    "ruff>=0.8.0",
    "black>=24.0",
    "pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
